requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-braces==1.14.0
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-haystack==2.8.1
idna==2.9
Pillow==7.1.1
pysolr==3.9.0
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
